SPR-13735 - Move ChainedTransactionManager from spring-data commons to spring-tx - fix issues found by codacy
